+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Fix `maxAge` of the `Stream.buffer` operator.
 * Add `Fraction.sign` and `Complex.sign`.
 * Add `IterablePrinter` to print lists.
+* Add `Ordering.percentile`. 
+* Relax the type constraints on `Ordering.natural`.
 * `CharMatcher` is now a `Pattern`.
 * Add `package:more/feature.dart`.
 * Add `Trie` (prefix tree) collection type.